Moved everything to the CortanaFunctions class.
Also made it so you can define the C# code for each Command Name inline in the vcdLookup dictionary, not as individual functions.

Also, added two examples, one for opening a website and another for opening the a file.

Still need to figure out if its possible to open a program or run a script.